+++ packages/trading-view/src/tradingView/tradingViewAdapter/hooks/useCreateRenderer.ts
@@ -1,10 +1,14 @@
-import {useRef, useEffect, useContext} from 'react';
+import {useRef, useEffect, useState} from 'react';
 import {Renderer} from '../renderer/renderer';
 import {useOrderStream, usePositionStream} from '@orderly.network/hooks';
 import {OrderStatus} from "@orderly.network/types";
+import { TpslAlgoType } from '../renderer/tpsl.util';
+import { AlgoType } from '../type';
+import { DisplayControlSettingInterface } from '@/tradingView/tradingView';
 
-export default function useCreateRenderer(symbol: string) {
-    const renderer = useRef<Renderer>();
+export default function useCreateRenderer(symbol: string, displayControlSetting?: DisplayControlSettingInterface) {
+    const [renderer, setRenderer] = useState<Renderer>();
+    const rendererRef = useRef<Renderer>();
 
     const [{rows: positions}, positionsInfo] = usePositionStream(symbol);
     const [pendingOrders] =
@@ -15,13 +19,23 @@ export default function useCreateRenderer(symbol: string) {
 
     const createRenderer = useRef(
         (instance: any, host: any, broker: any) => {
-            renderer.current = new Renderer(instance, host, broker);
-            renderer.current = new Renderer(instance, host, broker);
+            if (rendererRef.current) {
+                rendererRef.current.remove();
+            }
+            rendererRef.current = new Renderer(instance, host, broker);
+            setRenderer(rendererRef.current);
         }
     );
-
+    const removeRenderer = useRef(() => {
+        rendererRef.current?.remove();
+        rendererRef.current = undefined;
+    });
 
     useEffect(() => {
+        if (!displayControlSetting || !displayControlSetting.position) {
+            renderer?.renderPositions([]);
+            return
+        }
         const positionList = (positions ?? []).map(item => {
             return {
                 symbol: item.symbol,
@@ -35,14 +49,65 @@ export default function useCreateRenderer(symbol: string) {
                 basePriceDecimal: 4,
             }
         });
-
-        renderer.current?.renderPositions(positionList);
-    }, [renderer.current, positions, symbol]);
+        renderer?.renderPositions(positionList);
+    }, [renderer, positions, symbol, displayControlSetting]);
 
     useEffect(() => {
-        renderer.current?.renderPendingOrders(pendingOrders);
+        let tpslOrder: any = [];
+        let positionTpsl: any = [];
+        let limitOrder: any = [];
+        let stopOrder: any = [];
 
-    }, [renderer.current, pendingOrders, symbol]);
+        const symbolPosition = ( positions ?? []).find(item => item.symbol === symbol);
+        pendingOrders?.forEach(order=> {
+            if (symbol !== order.symbol) {
+                return;
+            }
+            if (!order.algo_order_id) {
+                limitOrder.push(order);
+            }else if (order.algo_order_id) {
+                if (order.algo_type === AlgoType.POSITIONAL_TP_SL ) {
+                    for(const child_order of order.child_orders) {
+                        child_order.root_algo_order_algo_type = order.algo_type;
+                        if (child_order.trigger_price && child_order.status !== OrderStatus.FILLED) {
+                            positionTpsl.push(child_order);
+                        }
+                    }
+                } else if (order.algo_type === AlgoType.TP_SL) {
+                    if (symbolPosition) {
+                        for(const child_order of order.child_orders) {
+                            child_order.root_algo_order_algo_type = order.algo_type;
+                            child_order.position_qty = symbolPosition.position_qty;
+                            if (child_order.trigger_price && child_order.status !== OrderStatus.FILLED) {
+                                tpslOrder.push(child_order);
+                            }
+                        }
+                    }
 
-    return [renderer, createRenderer.current] as const;
+                } else if (order.algo_type === AlgoType.STOP_LOSS || order.algo_type === AlgoType.TAKE_PROFIT) {
+                    stopOrder.push(order);
+                }
+            }
+        });
+        if (displayControlSetting) {
+            if (!displayControlSetting.positionTpsl) {
+                positionTpsl = [];
+            }
+            if (!displayControlSetting.tpsl) {
+                tpslOrder = [];
+            }
+            if (!displayControlSetting.limitOrders) {
+                limitOrder = [];
+            }
+            if (!displayControlSetting.stopOrders) {
+                stopOrder = [];
+
+            }
+        }
+
+        renderer?.renderPendingOrders(tpslOrder.concat(positionTpsl).concat(limitOrder).concat(stopOrder));
+
+    }, [renderer, pendingOrders, symbol, displayControlSetting, positions]);
+
+    return [createRenderer.current, removeRenderer.current] as const;
 }