+++ apps/docs/CHANGELOG.md
@@ -1,5 +1,46 @@
 # @orderly/docs
 
+## 0.4.0
+
+### Minor Changes
+
+- dd94bfa9: update
+
+### Patch Changes
+
+- 6cb9c466: fix bug;
+- Updated dependencies [dd94bfa9]
+- Updated dependencies [6cb9c466]
+  - @orderly.network/react@1.5.0
+  - @orderly.network/hooks@1.5.0
+  - @orderly.network/web3-onboard@1.5.0
+  - @orderly.network/types@1.5.0
+
+## 0.4.0-alpha.1
+
+### Patch Changes
+
+- fix bug;
+- Updated dependencies
+  - @orderly.network/react@1.5.0-alpha.1
+  - @orderly.network/web3-onboard@1.5.0-alpha.1
+  - @orderly.network/hooks@1.5.0-alpha.1
+  - @orderly.network/types@1.5.0-alpha.1
+
+## 0.4.0-alpha.0
+
+### Minor Changes
+
+- update
+
+### Patch Changes
+
+- Updated dependencies
+  - @orderly.network/react@1.5.0-alpha.0
+  - @orderly.network/hooks@1.5.0-alpha.0
+  - @orderly.network/web3-onboard@1.5.0-alpha.0
+  - @orderly.network/types@1.5.0-alpha.0
+
 ## 0.3.6
 
 ### Patch Changes