+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @orderly.network/cli
 
+## 0.3.0
+
+### Minor Changes
+
+- dd94bfa9: update
+
+### Patch Changes
+
+- 6cb9c466: fix bug;
+- Updated dependencies [dd94bfa9]
+- Updated dependencies [6cb9c466]
+  - @orderly.network/core@1.5.0
+
+## 0.3.0-alpha.1
+
+### Patch Changes
+
+- fix bug;
+- Updated dependencies
+  - @orderly.network/core@1.5.0-alpha.1
+
+## 0.3.0-alpha.0
+
+### Minor Changes
+
+- update
+
+### Patch Changes
+
+- Updated dependencies
+  - @orderly.network/core@1.5.0-alpha.0
+
 ## 0.2.6
 
 ### Patch Changes