+++ packages/component/CHANGELOG.md
@@ -1,5 +1,52 @@
 # @orderly.network/react
 
+## 1.5.0
+
+### Minor Changes
+
+- dd94bfa9: update
+
+### Patch Changes
+
+- 6cb9c466: fix bug;
+- Updated dependencies [dd94bfa9]
+- Updated dependencies [6cb9c466]
+  - @orderly.network/trading-view@1.5.0
+  - @orderly.network/core@1.5.0
+  - @orderly.network/hooks@1.5.0
+  - @orderly.network/net@1.5.0
+  - @orderly.network/types@1.5.0
+  - @orderly.network/utils@1.5.0
+
+## 1.5.0-alpha.1
+
+### Patch Changes
+
+- fix bug;
+- Updated dependencies
+  - @orderly.network/trading-view@1.5.0-alpha.1
+  - @orderly.network/hooks@1.5.0-alpha.1
+  - @orderly.network/types@1.5.0-alpha.1
+  - @orderly.network/utils@1.5.0-alpha.1
+  - @orderly.network/core@1.5.0-alpha.1
+  - @orderly.network/net@1.5.0-alpha.1
+
+## 1.5.0-alpha.0
+
+### Minor Changes
+
+- update
+
+### Patch Changes
+
+- Updated dependencies
+  - @orderly.network/trading-view@1.5.0-alpha.0
+  - @orderly.network/core@1.5.0-alpha.0
+  - @orderly.network/hooks@1.5.0-alpha.0
+  - @orderly.network/net@1.5.0-alpha.0
+  - @orderly.network/types@1.5.0-alpha.0
+  - @orderly.network/utils@1.5.0-alpha.0
+
 ## 1.4.3
 
 ### Patch Changes