+++ packages/types/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @orderly.network/types
 
+## 1.5.0
+
+### Minor Changes
+
+- dd94bfa9: update
+
+### Patch Changes
+
+- 6cb9c466: fix bug;
+
+## 1.5.0-alpha.1
+
+### Patch Changes
+
+- fix bug;
+
+## 1.5.0-alpha.0
+
+### Minor Changes
+
+- update
+
 ## 1.4.3
 
 ### Patch Changes